Deduplicate JSON path codec into a shared module.
Replace twin quote/escape scanners in file-viewer and JsonTree with one
format/parse/resolve implementation so arbitrary keys stay round-trippable
without duplicated path logic.
